Calculate next-boot CPU features on-the-fly
Calculate the featureset that the VM requires when it (re)boots each
time when pushing new metadata to the VM. The calculation depends on the
pool's feature level and the VM's domain-type, where the latter is also
applies on next-boot. A new metadata push is triggered automatically
when the domain-type is changed. Altogether, this ensures that the
domain-type and featureset are always in sync when a VM (re)boots.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
(cherry picked from commit b2df6fee0324635e7fd8b630c26c01656d514cbe)