CP-38850 add xapi.conf option for cert-expiration-days
When xapi generates/refreshes a host or pool certificate, the number of
days it will remain valid will be taken from xapi.conf.

Note that initially certficates are generated using a binary 'gencert'
which does not read this value from xapi.conf and will use the current
default of 10 years.

Also note that, for now, certificates used by clusterd have an unlimited
lifetime but this could be changed to also use the value specified in
xapi.conf.

To generate new certificates with a lifetime as in xapi.conf:

* edit xapi.conf
* restart xapi to make it read xapi.conf: service xapi restart
* xe host-reset-server-certificate
* xe host-refresh-server-certificate host=<hostname>

To verify SSL certficate lifetimes:

* openssl s_client -showcerts -connect localhost:443  2> /dev/null < /dev/null
    | openssl x509 -noout -dates

* openssl s_client -showcerts -connect localhost:443
    -servername pool 2> /dev/null < /dev/null | openssl x509 -noout -dates

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>