maintenance: remove warnings
Both xapi-types and xapi-client generate code, so it's easier to ignore
the warnings here rather than change the generators.

Warning 33 is ignored here because `deriving rpc` opens the Rpc module
even if it's not used.

Warning 50 is emitted because of empty docstrings.

After these warnings are removed a whole lot of warnings are uncovered
when running `dune build`, some of them are worth looking at as they
ignore types returned.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>