REQ-403 add cert checking for clusterd
Certificate checking for clusterd is implemented by relying on the
certificates xapi uses for intra-pool communication. This implies that
only hosts part of a xapi pool can form a cluster. Each host in the pool
has a certficate (public and private key) and knows the public keys of
all hosts in the pool. These are kept in PEM bundles.

We remove code from a previous design where clusterd used a single
certificate generated by xapi which was passed to clusterd.

A tls_config, sent to clusterd, contains the essential information:

* The common name (CN)
* Path the server certificate
* Optional path to bundle with trusted certificate or None if clusterd
  should not perform certficate checking.

The tls_config is global per cluster and hence the CN has to be a global
value and can't be specific per host. For now, use a simple string to
avoid confusion.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>