Automatically record the versions of new datamodel elements
A new program detects datamodel elements that do not yet have any
release version assigned to them. It then generates a file at
`ocaml/idl/datamodel_lifecycle.ml` which is used to supplement the
lifecycle metadata of these elements.

We are moving away from named releases in the datamodel to the more
finegrained notion of the xapi version, defined by git tags.
Automatically recording versions rather than having to manually annotate
the datamodel helps to avoid mistakes and omissions.

When a new class, field or message is added to the datamodel, it is not
yet known in which the software version it will be released, as the
version tag will be added later. A temporary version with suffix -next
is therefore first added to newly discovered elements. After a
version has been assigned, this replaces the temporary ones.

The system sets the Prototyped field in the datamodel lifecycle rather
than the Published field, to allow for features that are initially
experimental and their API not yet stable. Once datamodel entries are
declared stable, a separate process will have to mark them as Published.
We can provide automation to help with this as well.

In short, the workflow for adding a new element to the datamodel is as
follows:

- Add the new class/field/message with `~lifecycle:[]`.
- Run make, which will update `ocaml/idl/datamodel_lifecycle.ml` to record
  the new element with `<current version>-next`.
- Commit and raise a PR as usual.
- At some point later, a new version tag will be added.
- After that, whoever does a build will see another update of
  `ocaml/idl/datamodel_lifecycle.ml`, changing the *-next versions to
  the new version. This must be committed as part of the following PR.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>