REQ-403 gencert.service - generate pool cert first
Certificates for xapi's API clients and internal pool communication are
generated by gencert.service, which invokes gencert for this. The
certificate for the API clients requires the IP of the management
interface (obtained by gencert), which might not be yet available and
causes this to fail. The certificate for the pool communication does not
suffer from this dependency. So generate it first to have it available.
If the second call then fails, the systemd will run it again. But
gencert is idempotent such that the already created certificate for pool
communication won't be overwritten.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>