maintenance: change wlan.py to act like wake-on-lan.py
This allows wlan.py to act as both part of the power-on functionality as
well as a standalone script.

The wake_on_lan function's signature is unchanged as it's used in other
python modules. In practice it means that now the preliminary actions
that were done in wake-on-lan's main function are done in wlan's
wake_on_lan

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>