CP-39031 keep more xapi version details for Host.software_versions
Keep more details about the version of xapi as it is recorded in the
"sofware-versions" field of a host. This is most useful on a host where
a new version has been installed but not yet started. We want to know,
what version is currently running. To avoid problems with changing the
format of "software_version", add this under a new field "xapi_build".

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>