CA-363633: Always take the generation-id directly from xapi
The generation-id of a VM is made up by xapi - specifically for Windows
VMs - and xapi changes it or leaves it the same depending on the use
case. One such use case is revert-from-checkpoint, where the reverted VM
is expected to get a new generation-id.

The generation-id is put into the plaformdata by xapi and xenopsd
just writes it to xenstore when creating a domain. Because of a recent
change, the platformdata is now persisted by xenopsd across
suspend/resume and migration, and will therefore not take any new values
that set by xapi in the VM metadata.

Since a checkpoint is essentially a suspended VM, and a
revert-from-checkpoint results in a VM that can subsequently be resumed,
the VM now does not get a refreshed generation-id after this operation,
which is a bug.

To address this, the generation_id is given a dedicated field in the VM
metadata for xenopsd, so that it behaves as before. Xenopsd itself then
puts it into the platformdata when creating a domain.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
(partially cherry picked from commit ff3b76a7b9459a4a714996ea533d8c5710a93e73)