CA-364450: Fix YUM repo config for repo metadata checking
The '--setopt' of yum-config-manager requires specifying the repo name.
The bug fixed in this commit is '--setopt=repo_gpgcheck=1' should be
'--setopt=<repo-name>.repo_gpgcheck=1' in parameter of
'yum-config-manager'.

In this commit, additionally the `repo_gpgcheck` is moved to the initial
repo configuration file content which contains static configurations
usually.

Signed-off-by: Ming Lu <ming.lu@citrix.com>