Fix random insert of fist point
The cert distribution code provides a fist point for testing. The fist
point takes a list of operations and inserts randomly a failure. We have
seen logs where that failure was inserted at the end - which does not
make sense for testing. Replace the code that inserts the operations to
make sure it never inserts the failure at the end unless the list is
empty.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>