CP-39551: avoid warnings in xapi-cli-server
There were a lot of warnings here, mostly missing named parameters,
quite a few of these were able to be automated using sed to add tildes,
others couldn't because of how irregular the name of the parameters
are, sometimes using self, sometimes the type of object; or they were
too particular.

There we also unused bindings which allowed for removal of parameters.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>