CP-32574: Apply livepatches
This commit adds functions to apply live patches for two components:
Xen hypervisor and Linux kernel. A live patch can make changes on the
running component but without host reboot.

A live patch file is delivered by a RPM package. After that, Toolstack
will determine if an installed (or to be installed) live patch is
applicable to the current running component. If a live patch is
applicable, it will be applied by Toolstack.

A live patch has its own guidance which will overwrite the guidance in
the its component's RPM update.

The failures of live patches will not block/fail the update. But these
failures will be returned from 'host.apply_updates'.

Signed-off-by: Ming Lu <ming.lu@citrix.com>