CA-365121: pool join: require common xapi versions
In the pre-join checks, require that the xapi version of the joining
host is equal to that of the pool coordinator. This is stricter than,
and replaces, the check for common platform versions. In particular,
this protects against joining a host with a newer xapi version to a
pool, which may break (the coordinator must at any time have the highest
version).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>