CP-39441: Enable xenopsd to directly wait on pid creation
Xenopsd manages several services that sustain the lifecycle of each
domain. To launch a service, xenopsd usually starts up a daemon and it
waits for this daemon to write its pid on xenstore to signal the service
is ready to attend requests.

While this method is convenient in order to be able to cancel the wait
using xenstore watches, it's unfortunate as it makes the daemons depend
on xenstore even if it isn't needed to supply the service.

This change implements an alternative way to set up a cancellable wait
on the service readiness using inotify and epoll by setting up
filesystem watches instead of xenstore ones.

The polling loop wakes up every second to have a direct method to deal
with timeouts as well as unrelated events that happened in the
filesystem.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>