CP-39134: varstore-guard: use inotify to wait for the apperance of the socket
This is the socket created by xapi-guard itself through Conduit,
but Conduit doesn't currently have an API to notify us when the socket
got created and listen()-ed on.
And we don't want to return too early to our caller
(the unit test or xenopsd) since they may get an ECONNREFUSED if they
try to connect too early.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>