CA-366098: Raise internal xenopsd error on task timeout
Previously the task was reported as cancelled, which was nonsensical and
made triaging issues harder than necessary. Instead report an internal
error as this is not expected. The task is still cancelled to make it
disappear.

The internal wait method has been changed to be able to return relevant
information about the event captured, allowing to not just return a
boolean if desired. This is useful to retrieve information about tasks
just once instead of losing the information and making callers retrieve
again the information which makes the code flow difficult to navigate
and understand.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>