Add binary xapi_gzip for testing Xapi_compression
The compression library Xapi_compression spawns processes to compress
and uncompress streams. To facilitate easier testing, add a small
stand-alone filter that compresses or uncompresses its standard input.
This filter requires the presence of forkexecd to work - so typically a
Citrix Hypervisor installation. We don't install this binary but it is
available for developers.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>