Add temporary feature 'Internal_repo_access' to allow update in mix mode
In commit d951478f3, a new function was added to restrict the repository
access to pool members only. With this funciton, the pool members need
to use pool_secret in access requests to the mirrorred repositories on
the pool coordinator for authentication.

But in updating, this would cause the pool members can't access the
mirrored repositories as at that time the pool coordinator has been
updated to check the pool_secret while the pool members have not been
udpated to use pool_secret.

Per suggestion from Rob Hoes, this commit adds a temporary feature flag
to make the pool coordinator be able to distinguish the case in the
mixed mode in updating.

Signed-off-by: Ming Lu <ming.lu@citrix.com>