Fixup: Reverse the bool value if RPM package updates available
This commit fixes the bug introduced in c0814421f:
prior to c0814421f, the bool value is for [up_to_date], while after the
commit, it is for [pkg_updates_available].
If there is(are) RPM package updates, it means a host is NOT up to date.

This commit is just for this bug.

Signed-off-by: Ming Lu <ming.lu@citrix.com>