Use file type for is_raw_image()
When resuming a VM, xenopsd branches based on where the suspend image is
stored. So far it tries to open the image and read the header. This
either succeeds, in which case it concludes this is a raw image, or
fails. It then open the image again to read it. This patch simplifies
this logic by using the file type to make that decision. If the header
later can't be read this still would lead to a failure - so we are not
skipping the header check.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>