CP-39640/CP-39157 Add stream compression for VM migration
Optionally compress the memory byte stream that is coming out of
emu-manager as part of a VM migration and extend API and CLI for that
option. By default compression is not enabled.

Compression is implemented by feeding the byte stream through a
compression binary that we spawn with Forkexed. Currently we are using
zstd to achieve high troughput. Compression appears only benefical if
otherwise the network connection becomes saturated which often will
require multiple migrations in parallel.

At the API/CLI level compression is enabled by using the option
"compress=true"

The sending side signals to the receiving side the use of compression by
using a HTTP cookie key-value pair with "compress=true". If compression
is used, the handshaking algorithms is modified to not use a final
handshake. This is necessary to avoid a deadlock: the receiving side
sends the final handshake once all data is received but the sending side
can only do this (with a compressor in the data pipeline) by closing the
connection to force the compressor to process all data. Hence, it can't
wait for that final handshake before closing the connection.

In xenopsd the representation of VM_receive_memory operation is changed
from a tuple to a record because it can't be extended beyond 4 tupes to
include the new vmr_compressed flag. It signals if compression is used.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>