CP-39640/CP-39157 Add stream compression for VM migration
Optionally compress the memory byte stream that is coming out of
emu-manager as part of a VM migration and extend API and CLI for that
option. The default (when no option is used) is controlled via xapi.conf
and the current default is to not use compression. When switched on in
xapi.conf and no explicit API/CLI option is provided, localhost
migration is not using compression because it does not add a benefit.

Compression is implemented by feeding the byte stream through a
compression binary that we spawn with Forkexed. Currently we are using
zstd to achieve high troughput. Compression is benefical if otherwise
the network connection becomes saturated which often will require
multiple migrations in parallel.

At the API/CLI level compression is enabled by using the option
"compress=true", disabled by "compression=false" and using the default
(controlled via xapi.conf) otherwise.

The sending side signals to the receiving side the use of compression by
using a HTTP cookie key-value pair with "compress=true".

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>