CA-366801 UPD-825 xsh: fix XAPI blob sync and EBADF
Backport of 11cce30639d84a8ffc7d3f9ddc5c5b23d0f9d018

xsh was closing the file descriptor both in the 'proxy' function and in
the stunnel wrapper code.  Push down the Unixfd.(!) dereference into the
proxy code to make it obvious that we have a Unix.close there, and
replace that with Unixfd.safe_close which avoids double close.

This was causing blob sync to fail in XAPI with an rsync protocol error
(due to the extra EBADF printed by xsh).

Fixes: c6511f0a48e1c058dd43bb02fc9503be0c1aca42 "CA-337546: update to new Stunnel_cache API with Safe_resources support"

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>