CA-367120: Add un-installed packages into accumulative update list
This commit fixes a bug in following case:
One new RPM package is to be installed due to being required by an
existing package. The new package has not been installed. There are
accumulative updates for this new package. Example:
foo-1-1.rpm introduced in update-01, but not installed
foo-2-1.rpm introduced in update-02, but not installed
foo-3-1.rpm introduced in update-03, and to be installed in this update

In this case, the accumulative updates (update-01 and update-02) should
be included in consolidating update information.

Signed-off-by: Ming Lu <ming.lu@citrix.com>