CP-38688 introduce Message.destroy_many() API/CLI call
The RPC overhead when destroying many messages is high. To avoid it,
introcuce a message to destroy several messages in one call. This
introduces a new API call Message.destroy_many. The corresponding CLI
command message-destroy is extended to accept a parameter "uuids" to
accept multiple UUIDs in additio to the existing "uuid" parameter.

I notice that destroyig a message by itself is not cheap because it
parses the XML content of the message being destroyed. In an unlucky
case it also involves scanning a directory on the file system to remove
an alias.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>