CP-39136: Reintroduce backend to vtpms
This type it's for signifying what persistence backend it's being used.
This means its type from a vm ref to an enum, for now with a single
possible value. There's no way to change yet backend, I would be wary of
adding them, as this is intended to permit several persistence backends
at the same time only for the benefit of developers.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>