CP-39136: Reintroduce backends for VTPMs
Reintroduces the previous "backend" field which is unused, kept to avoid
issues with updating the schema database, now it's DynamicRO as it's not
used in the constructor.
Introduced the persistence backend field to declare how is the VTPM
stored. For now there's only a single possible value: xapi's DB.
This is intended to permit several persistence backends at the same time
only for the benefit of developers and it's not meant to allow changing
the backend for any single VTPM.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>