Upgrade VM runtime state when xenopsd restarts
This is needed for the case that the type of the backend's internal
state is extended in a xenopsd update, such that any running VMs have
the current runtime metadata. Case in point is the recent addition of
the platformdata key, where the default value (an empty map) is not
sufficient.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>