XSI-1246/CA-367232: Daily license re-apply fails is HA is enabled
A CH host is meant to re-apply it's license once a day, in order to
correctly handle expiry and to pick up any new licenses. This is some by
setting a timer in a thread in v6d, which triggers a host.apply_edition
call to xapi after 24 hours. The problem is that this call is rejected
if HA is enabled, which breaks the mechanism.

We'll fix this by allowing a host.apply_edition call if the edition is
not being changed. This is still consistent with the original intention
of the restriction (as evidenced by the code comment).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>