CP-39894: move vgpu starting code to service module
This allows to hide the arg-building into the module, as well as
removing wait_path from the interface.

Not all starting code can be moved, as this creates a module dependency
loop because of the use of the PCI module.

This means that the usage of the xenstore key as a cancellable watch has
to leak through the interface.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>