CP-39894: Replace is_pidfile and pid_path with pid_location
This forces the daemon modules to define which of the two paths they use
to advertise availability.

The translation is as follows:
use_pidfile = false means that only xenstore is used
use_pidfile = true means that both xenstore and the filesystem are used,
depending on situation.

Note that this exposes a current issue in swtpm: it is forced to define
a xenstore path when it is not used at all and the filepath is hardcoded
in the waiting function.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>