CP-39894: tweak Service.Qemu interface
This allows for clearer meaning in the paths offered by the module, as
well as removing the xenstore path by integrating the stop code in it
and simplify building of the arguments, which hasn't been pulled into the
module due to complexity.

The Vusb code has been simplified, it was checking twice the pid of
Qemu, this removes a race condition that could lead to a spurious error.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>