CP-39894: Use pid_location for file and xenstore cleanups
This allows for clearer logic regarding which conditions trigger each
cleanup.

Also quite a bit of code from the functions can be removed as the client
modules take control of the code.

Prevent error messages when trying to shutdown swtpm and varstored for a
domain that does not need them so they are not running. Instead print an
informational message about not bothering to shut it down.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>