fixup! CP-38583 add Host.last_software_update field with data/time
We want to use Xapi_host.get_servertime instead of Date.localtime
because it impacts the representation used for XMLRPC. This is obviously
a design deficiency as the representation of dates in XMLRPC should be
decided during serialisation and not at creation.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>

fixup! CP-38583 add Host.last_software_update field with data/time

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>