Use XMLRPC date format for Host.software_version.date
Change the date format for the build as it is recorded in the
Host.software_version.date field to match the XMLRPC date format for
consistency and expectations by Citrix Hypervisor Center.

The format used by XMLRPC http://xmlrpc.com/spec.md
is claimed to be ISO8601:

  19980717T14:08:55

This is somewhat doubtful becuase it mixes no punctuation for the date
with punctuation for the time.

https://stackoverflow.com/a/36775198

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>