scripts/xentrace: detect host CPU spikes and dump xentrace
Run xentrace in continuous tracing mode to a circular buffer
(of fixed size).
When a CPU spike is detected xentrace is sent a SIGTERM with `killall`
to get it to dump the xentrace buffer to disk.
The trace is compressed and another xentrace started.

This repeats until we run out of the space on the VDI set up for
tracing.

CPU spikes are detected by using `rrd2csv` and watching the host
`cpu_avg` which should be the average of the CPU usage of all the host's
pCPUs.

New flags:
* -c to enable circular tracing mode (default is a -T time one-off trace)
* -p <cpu-usage> e.g. -p 98 the detection point for CPU spike
* -r <n> repeat count - how many high cpu usage measurements in a row
  before triggerring the xentrace dump
* -M <n> to set the size of the circular buffer in MiB (default: 400MiB)

There is also a (hardcoded for now) sleep 60 between the traces just
so we don't continuously measure a very long spike.

To stop circular tracing mode you have to Ctrl-C on the running
xe-xentrace (e.g. under `screen`), or `killall xe-xentrace`
(but this seems to prevent automatic cleanup)

Signed-off-by: Edwin Török <edwin@etorok.net>