vTPM: add minimal support for saving/restoring state through the XAPI DB
On VM start read the vTPM state from the XAPI DB and write it out to a
file that is passed as argument to `swtpm-wrapper`.
On VM stop read the vTPM state from the filesystem and save it back into
the XAPI DB.

Note: any updated to vTPM state inbetween start/stop are lost if the
host running the VM crashes for now. To be addressed by other storage backends.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>