vTPM: add minimal migration support
Write a new swtpm record (similar to varstore record) into the migration
stream that contains the latest vTPM state (from the filesystem).
When vTPM start is called on the other end there will be 2 sources of
vTPM state: the migration state and the xapi DB. The data from the
migration state is restored first and takes precedence: there is a check
to guard against EEXIST when starting the vTPM: if it already exists
then we must've been just migrated and we leave the state alone.

This assumes that state would be stored in the filesystem (through
either the file:// or dir:// backends of swtpm, currently the only 2
existing ones).
If other backends are implemented in the future then we would need to
retrieve the latest vTPM state via another mechanism.

TBC: would qemu attempt to restore the device state as well? But we need
to start swtpm with *some* state during migration...

Signed-off-by: Edwin Török <edvin.torok@citrix.com>