CA-367979: Return changed guidance from host.apply_updates
Prior to this commit, the update won't change the guidances after
updating on a host. On failures of updating, no guidances will be applied
on the host. With supporting livepatch function, however, this doesn't
hold anymore. The failure of applying a livepatch may change the
guidances. THe final guidances should be returned to XAPI clients.

This commit addes a function to return the changed guidances if there are
changes.

Signed-off-by: Ming Lu <ming.lu@citrix.com>