CA-368231: Require only a VM reference for vtpm-create
We don't want users of the constructor to have full control of the
record, in particular control over the contents. This is because this
leaks the concrete storage backend, which makes this very brittle.
Instead limit the constructor to take only a VM reference.

In the future the constructor we probably want to change the constructor
to accept the profile the VTPM needs to have.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>