Maintenance: use dune-build-info
This will cause the version number used during builds and unit tests to
be 0.0.0, dune will only update the version in the installed executable.

To ensure the correct version is used in gen_lifecycle use binary
promotion (promote the .exe into the source dir, but don't commit it).

XAPI_VERSION can still be used to override the version for non-git
builds by appending it to `dune-project`. This will be used by .spec
builds.

Normally XAPI_VERSION will not be set at all: dune knows how to retrieve
it from git on its own.

Also handles the case where version numbers start with a `v`.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>