Maintenance: avoid rebuilds on date changes
By default on release builds inlining is performed: even if a module has
an interface file any changes to the .ml file will trigger recompilation
of all modules that use it.
For something like the version numbers or dates this could cascade
through a lot of modules.

Accessing this information is not performance critical: it is better to
avoid inlining and speed up the build by avoiding needless rebuilds.
It should also allow more builds to proceed in parallel.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>