CP-40392 compress vGPU migration stream
During VM migration memory and vGPU memory are streamed from source
to destination host. Currently, the memory stream can be compressed
optionally while the vGPU stream remains uncompressed. This patch adds
compression for the vGPU stream. It is controlled by the same mechanism
and as such is optional, too, and the default it controlled by a flag in
/etc/xapi.conf. Likewise, it us using the same compression framework
(but it would be easy to add one using different paramters).

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>