CA-365946: Block VIF and VBD hotplug into dom0
These operations fail anyway, because xenopsd doesn't manage dom0.
However, `currently_attached` was still erroneously set to `true`.

Unplug operations similarly fail, but we let them through and catch the
error in order to mark the device as not attached. This gives people a
way out if they had plugged a VIF or VBD by mistake, and they can't
get rid of it anymore.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>