CP-40087: replace vtpm profile with individual fields
is_unique is now required for creating vtpm objects. When creating VTPMs
from templates where it has the vtpm in platform data, is_unique:false
is used. This is because the mechanism is only needed for having a
backwards-compatible way to create VMs that can run windows 11.
Wanting to create unique VTPsM is not part of this so a new SDK will be
needed to use this feature.

This means removing the default vtpm profile from vms as well.

I've also removed some duplication that was present in xapi_xenops for
creating an empty vtpm.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>