CP-40155 Parallelize Host.evacuate
Host.evacuate migrates all VMs from a host to another in preparation of
a host update. This is currently done sequentially, one VM after
another. We want to migrate VMs in parallel to reduce total elapsed
time.

Because a host can host hundreds of VMs, we want to limit
parallelization such that at most n migrations are active at the same
time. We implement this using batching: a fixed number of VMs is
migrated off in parallel before the next batch is started. This is
controlled by a new parameter evacuation-batch-size in xapi.conf.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>