CA-368806: Workaround pbis get wedged
In this ticket, all PBIS lwsmd service threads keep in waiting
status and not able to service new requests. Just restart the
service can resovle the issue and bring the service back.

This symptom looks like a race condition. Given PBIS is going to
be replaced with winbind, this commit workaground the issue
instead of investigate and patch PBIS

* Restart lwsmd on LW_ERROR_ERRNO_ECONNREFUSED
* Restart lwsmd on xapi start/restart

Signed-off-by: Lin Liu <lin.liu@citrix.com>