CA-368437 remove duplicate keys from SM.features
The association list SM.features frequently contains duplicate keys.
This is problematic because they could bind to different values (which
are versions in this case).

When parsing the features from a list of strings, eliminate duplicate
keys. In addition, simplify the implementation and avoid duplicate work
by using List.filter_map.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>