CA-367236 replace Ezjsonm with Yojson
We would like to use just one library for creating and consuming JSON;
so replace Ezjsonm with Yojson.

The serialisation of timestamps in the V1 transport appears to change:
they are now serialised as floats; to avoid errors in the parse, we need
to change the reader for the V1 transport.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>