Simplify and fix UDSTransport implementation
The current implementation of UDSTransport in XenAPI.py and rrdd.py has
several issues:
- It redefines xmlrpclib.Transport.send_request, whose signature has changed
  in python 3, so this breaks XenAPI.py's python 3 port.
- The overridden implementation of send_request adds extra headers that are
  already handled by the parent class. As a result, extra headers are sent
  twice.
- Method __init__ is overridden needlessly, which could cause issues in
  the future if the parent class evolves.

This commit fixes the issues by simply deleting the send_request and
__init__ methods from the implementation of UDSTransport.

At the same time, code that was written to retain compatibility with python
< 2.7 is removed.

The add_extra_header method is kept in XenAPI.py, as `sm` apparently uses it
to add a `Subtask-of:` header. Symmetrically, I also kept it in rrdd.py,
but I'm not sure it's used anywhere.

Fixes #4767

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>