ci: work around dune version behaviour
Dune likes to pick the version from git over the declared version in
dune-project. Because on CI only the hash is available the versioning
schema cannot be followed.

Generating a valid version from this invalid version could be done, but
this would make catching honest mistakes when picking a version post-hoc
and more complex than needed.

Instead tag the current commit with a valid version so dune is forced to
do the right thing in a roundabout way.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>