CP-40444: Mark VTPM class as prototype
Although the module has existed since the Rio release, it doesn't make
sense to state so. This is because until now it was impossible to create
vtpm objects in the database, trying to create one of them resulted in
an exception. This has the added benefit to mark all the parts of the
class to be prototypes and this will make them easier to change before
the feature becomes stable.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>