CP-39744: simplify vm_platform.sanity_check parameters
The optional parameter for firmware served no purpose, and forced the
addition of a unit parameter, remove both.

This forces the unit-tests to use a particular firmware to run all
tests, use Bios as the default.

The firmware value is extracted in preparation of boot blocking
depending on the firmware and the presence of vTPM

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>