CP-40775: share function raising not done for vtpm exceptions
This makes it easy to track and link the removal of the function in the
quality gate to promotion of the feature to a stable one.

Also make raising the exception configurable so developers can easily
disable the check and test changes around the functionality without
having to recompile xapi. This is done by writing
xapiflags=-ignore-vtpm-unimplemented in /etc/sysconfig/xapi and
running xe-toolstack-restart

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>