Switch VM.pool_migrate over to HTTPS
This switches the xenopsd-to-xenopsd connection over to HTTPS, if
enabled in the config file (currently off by default).

Socket keepalives do not work when stunnel is used, as the given fd is
the local connection to stunnel, and are not set in HTTPS mode. This is
fine, because the stunnel client is already set up with keepalives.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>