Revert "Add a fallback system for auth files belonging to RPMs"
This reverts commit 5ad054a4baa39e0356e3c1001e2f557de3a56d50.

This commit caused xapi to delete dbx.auth from /usr/share/varstored.

Currently there's no way to recover this file once it's been deleted as
it only gets written on first boot.

There is a plan to make varstored read the certificates from /var/lib/
and make xapi write and delete the files there. While that is
implemented revert to the old behaviour to be able to brand shippable
versions.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>